fix: set token for `firmware/` requests from legacy devices
Legacy devices don't use the CoAP token, which is used here for caching. Without caching, block-wise transfers are far more demanding for CPU, memory and the HTTP server as every request is processed as a new one. For `firmware/` requests from legacy devices, indicated by having an 2109 header and no token, we create a token that is unique for that device and series of requests, the series being the block-wise transfer. This token is created using the SHA256 of the request path and header 2109.

Reference withthegrid/dtls-termination-proxy#35.